Fix where duplicated entries are being scraped from JMA
For some unknown reason, JMA's eq data have some duplicate entries upon checking on observed date of each entry. To fix this problem, I introduce a fix where if the date of the last checked entry is the same as the entry that is to yet to be added, then that will be skipped from being recorded. This fix is crucial as this can affect the accuracy of eq stats and other relevant analysis for these kinds of data.